+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 * Updated internal `AwaitTaskCorrect` implementation to align with [canonical version](http://www.fssnip.net/7Rc/title/AsyncAwaitTaskCorrect) [#49](https://github.com/fsprojects/FSharp.AWS.DynamoDB/pull/49) 
 * Added SourceLink info (using `DotNet.ReproducibleBuilds`)
 * Fixed `TableContext.UpdateTableIfRequiredAsync`: Guard against `NullReferenceException` when `StreamSpecification` is `null`
+* (breaking) Changed `TableContext.UpdateTableIfRequiredAsync`/`VerifyOrCreateTableAsync` to yield `TableDescription` (in order to surface ARNs)
 
 ### 0.11.0-beta
 * Added `Precondition.CheckFailed`